+++ packages/react-mui-shared/src/stories/PangeaFieldsForm/FolderFieldsForm.stories.tsx
@@ -39,6 +39,7 @@ export const FolderFieldsFormDemo: {
 
 FolderFieldsFormDemo.args = {
   object: {},
+  autoSave: true,
   fields: {
     general: {
       type: "grouping",