+++ packages/react-mui-store-file-viewer/src/components/ShareObject/utils.ts
@@ -1,4 +1,16 @@
+import { FC } from "react";
 import { ObjectStore } from "../../types";
+
+import LinkIcon from "@mui/icons-material/Link";
+import EmailIcon from "@mui/icons-material/Email";
+import SmartphoneIcon from "@mui/icons-material/Smartphone";
+import PasswordIcon from "@mui/icons-material/Password";
+
+import UploadIcon from "@mui/icons-material/Upload";
+import DownloadIcon from "@mui/icons-material/Download";
+import BorderColorIcon from "@mui/icons-material/BorderColor";
+
+import { IconProps, SvgIconProps } from "@mui/material";
 
 export const getShareDisplayName = (
   object: ObjectStore.ShareObjectResponse
@@ -12,7 +24,7 @@ export const getShareDisplayName = (
     identityAuth?.auth_type === ObjectStore.ShareAuthenticatorType.Email ||
     identityAuth?.auth_type === ObjectStore.ShareAuthenticatorType.Sms
   ) {
-    return `Secured with ${identityAuth.auth_context}`;
+    return `${identityAuth.auth_context}`;
   }
 
   if (identityAuth?.auth_type === ObjectStore.ShareAuthenticatorType.Password) {
@@ -20,6 +32,26 @@ export const getShareDisplayName = (
   }
 
   return `Shared using ${authCount} authenticators`;
+};
+
+export const getShareDisplayIcon = (
+  object: ObjectStore.ShareObjectResponse
+): FC<SvgIconProps> => {
+  if (!object) return LinkIcon;
+
+  if (object?.link_type === ObjectStore.ShareLinkType.Download) {
+    return DownloadIcon;
+  }
+
+  if (object?.link_type === ObjectStore.ShareLinkType.Upload) {
+    return UploadIcon;
+  }
+
+  if (object?.link_type === ObjectStore.ShareLinkType.Editor) {
+    return BorderColorIcon;
+  }
+
+  return LinkIcon;
 };
 
 export const getDateDisplayName = (dateString: string) => {