+++ packages/react-mui-store-file-viewer/src/stories/StoreEditorFileViewer.stories.tsx
@@ -95,14 +95,15 @@ StoreFileViewerDemo.args = {
           throw err;
         });
     },
-    upload: async (body) => {
+    upload: async (body, contentType) => {
       return axios
         .post(
           `https://store.${process.env.STORYBOOK_SERVICE_DOMAIN}/v1beta/put`,
-          { ...body },
+          body,
           {
             headers: {
               Authorization: `Bearer ${process.env.STORYBOOK_PANGEA_TOKEN}`,
+              "Content-Type": contentType,
             },
           }
         )