+++ packages/react-mui-store-file-viewer/src/types/index.ts
@@ -136,6 +136,9 @@ export namespace ObjectStore {
     max_access_count?: number;
 
     authenticators?: ShareAuthenticator[];
+
+    title?: string;
+    message?: string;
   }
 
   export interface ShareLinkToSend {
@@ -145,7 +148,8 @@ export namespace ObjectStore {
 
   export interface ShareSendRequest {
     links: ShareLinkToSend[];
-    from_prefix: string;
+    sender_email: string;
+    sender_name?: string;
   }
 
   export interface ShareSendResponse {
@@ -192,7 +196,7 @@ export namespace ObjectStore {
   export interface ShareObjectResponse {
     id: string;
     targets?: string[];
-    link_type?: string; // "upload" | "download";
+    link_type?: string; // "upload" | "download" | "editor";
     access_count?: number;
     max_access_count?: number;
 
@@ -216,6 +220,7 @@ export namespace ObjectStore {
   export enum ShareLinkType {
     Upload = "upload",
     Download = "download",
+    Editor = "editor",
   }
 
   export interface ShareListResponse {