+++ packages/react-mui-audit-log-viewer/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.58] - 2024-01-22
+
+### Fixed
+
+- Support sorting columns by all fields, except unindexed string fields
+
 ## [0.0.57] - 2024-01-10
 
 ### Changed