+++ packages/react-mui-shared/CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.43] - 2024-01-16
+
+### Added
+
+- Now export fields components:
+  - SelectField
+- PopoutCard component
+
+### Fixed
+
+- useInternalState not triggering callback function
+
+## [0.0.39] - 2024-01-12
+
+### Added
+
+- Now export fields components:
+  - AuthPasswordField, CheckboxField, DateTimeField, StringArrayField, JsonField, StringField, SwitchField
+- Multiline field supports TextFieldProps overrides and placeholder
+
 ## [0.0.37] - 2024-01-10
 
 ### Added