+++ packages/react-mui-store-file-viewer/CHANGELOG.md
@@ -5,6 +5,47 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.20] - 2024-01-23
+
+### Fixed
+
+- Updated share link confirmation button label to reflect selected authentication method
+
+## [0.0.19] - 2024-01-19
+
+### Fixed
+
+- Share link icons in file listing reflecting link type, support quick copy password post creation
+
+## [0.0.18] - 2024-01-19
+
+### Fixed
+
+- Download target to "zip" for DownloadFileViewer
+
+## [0.0.17] - 2024-01-16
+
+### Fixed
+
+- Added loading indicator to download search control
+- Removed uppercase on folder path
+
+## [0.0.14] - 2024-01-16
+
+### Added
+
+- Supporting passing title and message to share creation
+
+### Fixed
+
+- Updates share links creation modals
+
+## [0.0.10] - 2024-01-11
+
+### Fixed
+
+- Multi-select bulk deletion was not fully clearing deleted files. Added a delay to reload data, in case search results still contained recently delete files.
+
 ## [0.0.9] - 2024-01-10
 
 ### Fixed